fix test build error

+++ src/include/sys/types.h
@@ -1,6 +1,6 @@
 #ifndef _TYPES_H
 #define _TYPES_H
 
-typedef int ssize_t; 
+typedef long int ssize_t; 
 
 #endif
